+++ R/backend-mssql.R
@@ -469,7 +469,13 @@ simulate_mssql <- function(version = "15.0") {
       },
       all = mssql_bit_int_bit(win_aggregate("MIN")),
       any = mssql_bit_int_bit(win_aggregate("MAX")),
-      row_number = win_rank("ROW_NUMBER", empty_order = TRUE)
+      row_number = win_rank("ROW_NUMBER", empty_order = TRUE),
+
+      n_distinct = function(x) {
+        cli_abort(
+          "No translation available in `mutate()`/`filter()` for SQL server."
+        )
+      }
     )
 
   )}