+++ NEWS.md
@@ -1,5 +1,8 @@
 # dbplyr (development version)
 
+* SQL server: clear error if you attempt to use `n_distinct()` in `mutate()`
+  or `filter()` (#1366).
+
 * Add translations for clock functions `add_years()`, `add_days()`, 
   `date_build()`, `get_year()`, `get_month()`, `get_day()`, 
   and `base::difftime()` on SQL server, Redshift, Snowflake, and Postgres.